Fixed some bugs

+++ 03.Operators-Аnd-Expressions/09.Trapezoids/Trapezoids.cs
@@ -12,11 +12,11 @@ class Trapezoids
     /// </summary>
     static void Main()
     {
-        float a = float.Parse(Console.ReadLine());
-        float b = float.Parse(Console.ReadLine());
-        float h = float.Parse(Console.ReadLine());
+        double a = double.Parse(Console.ReadLine());
+        double b = double.Parse(Console.ReadLine());
+        double h = double.Parse(Console.ReadLine());
 
-        float area = ((a + b) / 2) * h;
+        double area = ((a + b) / 2) * h;
         Console.WriteLine("{0:F7}",area);
     }
 }
