+++ 03.Operators-Аnd-Expressions/12.N-thBit/NthBit.cs
@@ -11,16 +11,15 @@ class NthBit
     /// </summary>
     static void Main()
     {
-        int p = int.Parse(Console.ReadLine());
-        int n = int.Parse(Console.ReadLine());
+        long p = long.Parse(Console.ReadLine());
+        byte n = byte.Parse(Console.ReadLine());
 
         if (n < 55 && p >= 0 && p <= Math.Pow(2,55))
         {
-            int mask = 1 << n;
-            int andMask = p & mask;
-            int bit = andMask >> n;
+            long mask = 1 << n;
+            long andMask = p & mask;
+            long bit = andMask >> n;
             Console.WriteLine(bit);
         } 
     }
 }
-
